@leksss 2021-10-05
$ benchstat bench_old.txt bench_new.txt 
name              old time/op    new time/op    delta
GetDomainStat-12     216ms ± 1%       0ms ±10%  -100.00%  (p=0.008 n=5+5)

name              old alloc/op   new alloc/op   delta
GetDomainStat-12     133MB ± 0%       0MB ± 0%  -100.00%  (p=0.008 n=5+5)

name              old allocs/op  new allocs/op  delta
GetDomainStat-12     1.70M ± 0%     0.00M ± 0%      ~     (p=0.079 n=4+5)